requirements.txt
asgiref==3.9.1
beautifulsoup4==4.13.4
certifi==2025.8.3
charset-normalizer==3.4.3
cloudinary==1.44.1
colorama==0.4.6
distlib==0.4.0
Django==5.2.4
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-filter==25.1
django-otp==1.6.1
django-storages==1.14.6
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
et_xmlfile==2.0.0
filelock==3.19.1
gunicorn==23.0.0
idna==3.10
numpy==2.3.2
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
pillow==11.3.0
platformdirs==4.3.8
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
pytz==2025.2
qrcode==8.2
razorpay==1.4.2
razorpay-py==0.0.2
reportlab==4.4.3
requests==2.32.4
setuptools==80.9.0
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
virtualenv==20.34.0
